Move guide and tests into components/optics/
