+++ acc/components/optics/guide.py
