+++ tests/components/optics/guide_cpu_instrument.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 import mcvine, mcvine.components as mc
-from mcvine.acc.components import guide
+from mcvine.acc.components.optics import guide
 from mcni import rng_seed
 def seed(): return 0
 rng_seed.seed = seed
