+++ tests/components/optics/mcvine_guide_cpu_instrument.py
